+++ v0/routes/users/confirmEmail.test.js
@@ -7,6 +7,7 @@ const { User } = require('../../../database/models')
 const { NotFoundError, UnauthorizedError } = require('../../errors')
 const { signJwt } = require('../../utils')
 const scopes = require('../../scopes')
+const { shouldNotAcceptInvalidToken } = require('./testUtils')
 
 const confirmEmailEndpoint = '/v0/users/:id/confirmEmail'
 
@@ -18,34 +19,10 @@ async function makeRequest(token, id) {
 	return req
 }
 
-test(`POST ${confirmEmailEndpoint} should not accept invalid auth token`, async t => {
-	const user = await User.create({
-		username: 'test',
-		password: '123456',
-		email: 'test@test.com',
-		emailConfirmed: false
-	})
-
-	const expiredToken = await signJwt(
-		// backdate token 2 seconds
-		{ iat: Math.floor(Date.now() / 1000) - 2 },
-		// set token to expire in 1 second
-		{ expiresIn: '1s' }
-	)
-	const invalidScopeTokens = await Promise.all([
-		signJwt({ scopes: [] }, { subject: `${user.id}`, expiresIn: '10h' }),
-		signJwt({ scopes: 'invalid' }, { subject: `${user.id}`, expiresIn: '10h' })
-	])
-	const invalidTokens = [null, 'invalid', expiredToken, ...invalidScopeTokens]
-
-	for (const token of invalidTokens) {
-		const { status, body } = await makeRequest(token, user.id)
-
-		t.is(status, UnauthorizedError.CODE, body.message)
-		t.is(body.code, UnauthorizedError.CODE)
-		t.truthy(body.message)
-	}
-})
+test(
+	`POST ${confirmEmailEndpoint} should not accept invalid auth token`,
+	shouldNotAcceptInvalidToken(makeRequest)
+)
 
 test(`POST ${confirmEmailEndpoint} should update confirmed email for user`, async t => {
 	const user = await User.create({
