+++ v0/routes/users/getUser.test.js
@@ -7,6 +7,7 @@ const { User } = require('../../../database/models')
 const { UnauthorizedError, NotFoundError } = require('../../errors')
 const { signJwt } = require('../../utils')
 const scopes = require('../../scopes')
+const { shouldNotAcceptInvalidToken } = require('./testUtils')
 
 const selfEndpoint = '/v0/users/:id'
 
@@ -18,34 +19,10 @@ async function makeRequest(token, id) {
 	return req
 }
 
-test(`GET ${selfEndpoint} should not accept valid auth token`, async t => {
-	const user = await User.create({
-		username: 'test',
-		password: '123456',
-		email: 'test@test.com',
-		emailConfirmed: false
-	})
-
-	const expiredToken = await signJwt(
-		// backdate token 2 seconds
-		{ iat: Math.floor(Date.now() / 1000) - 2 },
-		// set token to expire in 1 second
-		{ expiresIn: '1s' }
-	)
-	const invalidScopeTokens = await Promise.all([
-		signJwt({ scopes: [] }, { subject: `${user.id}`, expiresIn: '10h' }),
-		signJwt({ scopes: 'invalid' }, { subject: `${user.id}`, expiresIn: '10h' })
-	])
-	const invalidTokens = [null, 'invalid', expiredToken, ...invalidScopeTokens]
-
-	for (const token of invalidTokens) {
-		const { status, body } = await makeRequest(token, user.id)
-
-		t.is(status, UnauthorizedError.CODE, body.message)
-		t.is(body.code, UnauthorizedError.CODE)
-		t.truthy(body.message)
-	}
-})
+test(
+	`GET ${selfEndpoint} should not accept valid auth token`,
+	shouldNotAcceptInvalidToken(makeRequest)
+)
 
 test(`GET ${selfEndpoint} should get profile of authenticated user`, async t => {
 	const user = await User.create({
